requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
drf-spectacular==0.27.0
PyJWT==2.8.0
